version: refactor get_uname_info()
Some code from "builtin/bugreport.c" uses uname(2) to get system
information.

Let's refactor this code into a new get_uname_info() function, so
that we can reuse it in a following commit.

We may need to refactor this function in the future if an
`osVersion.format` config option is added, but for now we only
need it to accept a "full" flag that makes it switch between providing
full OS information and providing only the OS name.

Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Usman Akinyemi <usmanakinyemi202@gmail.com>