t5701: add setup test to remove side-effect dependency
Currently, the "test capability advertisement" test creates some
"expect" files which is used by other tests below it.

To remove the side-effect creates by this test, let's split up part of
it into a "setup"-type test which creates the "expect" files which gets
reused by multiple tests.  This will be useful in a following commit.

Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Usman Akinyemi <usmanakinyemi202@gmail.com>