builtin/update-server-info: remove the_repository global variable
Remove the_repository global variable in favor of the repository
argument that gets passed in "builtin/upload-server-info.c".

The RUN_SETUP macro is used in "git.c" when the 'update-server-info'
command is wired to the 'cmd_update_server_info()' function."
This means we can be sure that the `run_builtin()` function inside
"git.c" will always pass a valid `repo` variable to `cmd_update_server_info()`
when the `update-server-info` command is run inside a Git repository.

When the command is run outside a Git repository without the `-h`
option, the command will fail (`die`) inside the `run_builtin()` function
when the `setup_git_directory()` is called. So, the `cmd_update_server_info()`
would not be called at all. When `-h` is passed to the command outside a
Git repository, the `run_builtin()` will call the `cmd_update_server_info()`
function with `repo` set as NULL.

It is certain that the `update_server_info()` function would not be
called when the `repo` config is `NULL` since this only happens when the
`-h` option is used and the command would exit with code 129 before
getting to the `update_server_info()` function inside the
`usage_with_options()` function.

To prevent accessing a `NULL` value `repo`, it is necessary to check if
the `repo` has a valid value before calling the `repo_config` option
inside "update-server-info.c" since it comes before the
`usage_with_options()` function.

So, this change is safe and would not lead to any breakage.

Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Usman Akinyemi <usmanakinyemi202@gmail.com>